+++ src/application/controllers/pets/breeds/index.ts
@@ -1,0 +1,2 @@
+export * from './load-cat'
+export * from './load-dog'
